pyproject.toml
[project]
name = "gemini-tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
]

